seperate code, add feedback message
