+++ src/components/Search.jsx
@@ -1,0 +1,5 @@
+function Search() {
+return 
+}
+
+export default Search
